Move ground_truths into activities
