+++ oonipipeline/src/oonipipeline/temporal/activities/ground_truths.py
@@ -1,20 +1,18 @@
-import asyncio
 from dataclasses import dataclass
 import pathlib
 import logging
 
-from datetime import datetime, timedelta
+from datetime import datetime
 
 from temporalio import workflow, activity
 
 with workflow.unsafe.imports_passed_through():
     import clickhouse_driver
 
-    from oonidata.dataclient import date_interval
     from oonidata.datautils import PerfTimer
-    from ..analysis.control import WebGroundTruthDB, iter_web_ground_truths
-    from ..netinfo import NetinfoDB
-    from ..db.connections import (
+    from ...analysis.control import WebGroundTruthDB, iter_web_ground_truths
+    from ...netinfo import NetinfoDB
+    from ...db.connections import (
         ClickhouseConnection,
     )
 
@@ -60,31 +58,3 @@ def make_ground_truths_in_day(params: MakeGroundTruthsParams):
             rows=iter_web_ground_truths(db=db, measurement_day=day, netinfodb=netinfodb)
         )
         log.info(f"built ground truth DB {day} in {t.pretty}")
-
-
-@workflow.defn
-class GroundTruthsWorkflow:
-    @workflow.run
-    async def run(
-        self,
-        params: GroundTruthsWorkflowParams,
-    ):
-        task_list = []
-        start_day = datetime.strptime(params.start_day, "%Y-%m-%d").date()
-        end_day = datetime.strptime(params.end_day, "%Y-%m-%d").date()
-
-        async with asyncio.TaskGroup() as tg:
-            for day in date_interval(start_day, end_day):
-                task = tg.create_task(
-                    workflow.execute_activity(
-                        make_ground_truths_in_day,
-                        MakeGroundTruthsParams(
-                            clickhouse=params.clickhouse,
-                            data_dir=params.data_dir,
-                            day=day.strftime("%Y-%m-%d"),
-                            rebuild_ground_truths=True,
-                        ),
-                        start_to_close_timeout=timedelta(minutes=30),
-                    )
-                )
-                task_list.append(task)
